backend/requirements.txt
flask
flask_sqlalchemy
flask_cors
flask_jwt_extended
python-dotenv
psycopg2-binary
